==> experiment-1/modelling/08_size_hist_and_model.R <==
###### Frequency of rectangles by size #########
# This is a mess

rm(list = ls())
# setup F
library(here)
source(here("functions/getLearnerHypDistributions.R"))
source(here("functions/calculatingFunctions.R"))
source(here("functions/plottingFunctions.R"))

exp <- 1
alphas <- c(-1, 0, 1) # alphas we are interested in modelling
c <- 3 # clue we are interested in
b <- 8 # block we are interested in
priors <- "flat"
filter_cover_check <- 1
dif_priors <- FALSE 
if (length(priors) > 1) dif_prriors = TRUE


# load generic data
load(here(paste0("experiment-",exp,"/data/derived/all_conditions.Rdata")))
source(here("functions/getLearnerHypDistributions.R"))
load(here(
  "experiment-scenarios/target-blocks/data/target-block-8-Cartesian.Rdata"
))
observations = targetBlock$observations

# load participant data
load(here(paste0("experiment-",exp,"/data/derived/data_cartesian.Rdata")))

# filter participants by the number of times they passed the verification test if we're doing that
if(filter_cover_check){
  d_cartesian <- d_cartesian %>%
    filter(n_cover_check > filter_cover_check)
}

# Model predictions -------------------------------------------------------
# Predicted distributions for each alpha
all_dists = NULL
for (i in 1:length(alphas)) {
  dist <-
    getLearnerHypDistribution(observations, alpha = alphas[i], prior = "flat")
  # convert to df (probably a better way than this)
  distDf <- NULL
  # loop through clues if more than one clue to be plotted
  for (j in c) {
    distBlock <- dist[[j]]
    distDf <- rbind(distDf, distBlock)
    if (alphas[i] > 0) {
      distDf$cond <- "helpful"
    } else if (alphas[i] == 0) {
      distDf$cond <- "random"
    } else {
      distDf$cond <- "misleading"
    }
  }
  all_dists <- rbind(all_dists, distDf)
}

# uninformative condition (recursive learner, deception)
dist <-
  getLearnerHypDistribution(observations,
                            alpha = alphas[1],
                            prior = "flat",
                            recursion = TRUE)

# convert to DF
distDf <- NULL
for (j in c) {
  distBlock <- dist[[j]]
  distDf <- rbind(distDf, distBlock)
}
distDf$cond <- "uninformative"

# combine all conditions
all_dists <- rbind(all_dists, distDf)

# get probability distributions
all_dists <- all_dists %>%
  mutate(Alpha = as.factor(alpha)) %>%
  mutate(
    clue_name = case_when(
      clue == 1 ~ "1st Clue",
      clue == 2 ~ "2nd Clue",
      clue == 3 ~ "3rd Clue",
      clue == 4 ~ "4th Clue"
    )
  )


# Participant responses ---------------------------------------------------

# Need to make a single data frame with participant response, size, and probability.
# get probability of a certain sized rectangle under a certain alpha

d_sizes <- all_dists %>%
  group_by(cond, size) %>%
  summarise(prob = sum(posterior))%>% # using "sum" rather than the median or mean means that we can also consider how many rectangles there are in a given size category. 
  mutate(count = 0, Percent = 0)

# function to help match the frequency of rectangles generated by a participant with that rectangle's probability
replace_count <- function(d_size, d_cond) {
  for (i in 1:nrow(d_size)) {
    size_match <- d_cond$size_resp == d_size$size[i]
    if (any(size_match)) {
      d_size$Percent[i] <- d_cond$Percent[which(size_match)]
    }
  }
  return(d_size)
}

d_all_sizes <- NULL


all_conditions_filtered <- all_conditions %>%
  filter(clues == c & blocks == b)

# filter participants by the number of times they passed the verification test if we're doing that
if(filter_cover_check){
  all_conditions_filtered <- all_conditions_filtered %>%
    filter(conditions %in% c("HS", "MS", "RS","US"))
}

nConds <- length(all_conditions_filtered[, 1])



# Populate the "count" column in d_sizes for each condition, then make a single
# large data frame with sizes, probability, alpha, and frequency for each  condition
for (i in 1:nConds) {
  condition <- all_conditions_filtered[i, "conditions"]
  
  d_cond <- d_cartesian %>%
    filter(block == b & clue == c & cond == condition) %>%
    group_by(size_resp) %>%
    summarise(count = n()) %>%
    mutate(Percent = (count/sum(count))*100)
  
  # add column for experimental condition
  d_cond$cond <- condition
  
  # add column for cover story/model condition
  d_cond <- d_cond %>%   mutate(
    cover_cond = case_when(
      cond == "HS" | cond == "HN" ~ "helpful",
      cond == "MS" | cond == "MN" ~ "misleading",
      cond == "UN" | cond == "US" ~ "uninformative",
      cond == "RS" | cond == "RN" ~ "random"
    )
  )
  
  
  d_sizes_cover_cond <-
    d_sizes[d_sizes[, "cond"] == unique(d_cond$cover_cond), ]
  
  
  d_cond_sizes <- replace_count(d_sizes_cover_cond, d_cond)
  d_cond_sizes <- d_cond_sizes %>%
    mutate(cond = condition, prob = (prob*100)) # put prob on same scale as experiment data
  
  d_all_sizes <- rbind(d_all_sizes, d_cond_sizes)
}


file <-  paste0("experiment-", exp, "/data/derived/d_size_histograms-b",b,"-c",c)

if(filter_cover_check){
  file <- paste0(file,"-filter-cover")
}


save(d_all_sizes, file = here(paste0(file,".Rdata")))

# add column for cover story/model condition
all_conditions_filtered <- all_conditions_filtered %>%
  mutate(
    plotCond = case_when(
      conditions == "HS" | conditions == "HN" ~ "helpful",
      conditions == "MS" | conditions == "MN" ~ "misleading",
      conditions == "UN" | conditions == "US" ~ "uninformative",
      conditions == "RS" | conditions == "RN" ~ "random"
    )
  )

# Plot
# loop through all conditions and make a plot for each

for (i in 1:length(all_conditions_filtered[, 1])) {
  condition <- all_conditions_filtered[i, "conditions"]
  plotCond <- all_conditions_filtered[i, "plotCond"]
  sizeHistModel(d_all_sizes, condition, plotCond)
  
  
  file <-  paste0("experiment-", exp, "/data/derived/d_size_histograms-b",b,"-c",c)
  
  if(filter_cover_check){
    file <- paste0(file,"-filter-cover")
  }
  
  file <- paste0(
    "experiment-",exp,"/modelling/05_plots/size_hist_b8_c",
    c,
    "_",
    condition
  )
  
  if (filter_cover_check){
    file <- paste0(file,"-filter-cover")
  }
  
  
  ggsave(filename = here(
    paste0(
      file,".png"
    )
  ),
  width = 6,
  height = 4)
}
